AWS js sdk v3を利用するための関数をモジュール化
