+++ app/index.mjs
@@ -9,6 +9,7 @@ import {
   checkInvalidMailAddressFormat,
   checkVoidString,
 } from "./validation.mjs";
+import { confirmVerifiedAndSendEmailSes } from "./aws/SesModules.mjs";
 
 const createSendEmailCommand = (toAddress, fromAddress, contentObj) => {
   return new SendEmailCommand({
@@ -41,122 +42,6 @@ const createVerifyEmailIdentityCommand = (mailAddress) => {
   return new VerifyEmailIdentityCommand({ EmailAddress: mailAddress });
 };
 
-const verifyEmailAddressSes = async ({ sesClient, mailAddress }) => {
-  const verifyEmailIdentityCommand =
-    createVerifyEmailIdentityCommand(mailAddress);
-  console.log(mailAddress);
-  return sesClient
-    .send(verifyEmailIdentityCommand)
-    .then((res) => {
-      console.log("Success to verify email addmin");
-      console.log(res);
-      const successRes = {
-        statusCode: 200,
-        body: {
-          errorMessage: "メールアドレス検証成功",
-        },
-      };
-      const resJson = JSON.stringify(successRes);
-      return resJson;
-    })
-    .catch((err) => {
-      console.error("Failed to send email.");
-      console.error(err);
-      const awsError = {
-        statusCode: 500,
-        body: {
-          errorMessage:
-            "メールアドレス検証に問題がありました。Lambdaのログを確認してください。",
-        },
-      };
-      const resJson = JSON.stringify(awsError);
-      throw new Error(resJson);
-    });
-};
-
-const sendEmailSes = async ({ mailObject, sesClient, mailAddress }) => {
-  const sendEmailCommand = createSendEmailCommand(
-    mailAddress,
-    mailAddress,
-    mailObject
-  );
-
-  return sesClient
-    .send(sendEmailCommand)
-    .then((res) => {
-      console.log("Success to send email.");
-      console.log(res);
-      const successRes = {
-        statusCode: 200,
-        body: {
-          errorMessage: "メール送信成功",
-        },
-      };
-      const resJson = JSON.stringify(successRes);
-      return resJson;
-    })
-    .catch((err) => {
-      console.error("Failed to send email.");
-      console.error(err);
-      const awsError = {
-        statusCode: 500,
-        body: {
-          errorMessage:
-            "メール送信に問題がありました。Lambdaのログを確認してください。",
-        },
-      };
-      const resJson = JSON.stringify(awsError);
-      throw new Error(resJson);
-    });
-};
-
-const verifyAndSendEmailSes = ({ mailObject, ...sesSet }) => {
-  return verifyEmailAddressSes({
-    mailObject,
-    ...sesSet,
-  })
-    .then(() => {
-      return sendEmailSes({ mailObject, ...sesSet });
-    })
-    .catch((err) => {
-      console.error("Failed to send email.");
-      console.error(err);
-      return err;
-    });
-};
-
-const confirmVerifiedAndSendEmailSes = ({ mailObject, ...sesSet }) => {
-  const verifiedCheckInput = undefined;
-  const verifiedCheckCommand = new ListVerifiedEmailAddressesCommand(
-    verifiedCheckInput
-  );
-
-  return sesSet.sesClient
-    .send(verifiedCheckCommand)
-    .then((res) => {
-      const verifiedEmailList = res.VerifiedEmailAddresses;
-      return verifiedEmailList.includes(sesSet.mailAddress)
-        ? sendEmailSes({ mailObject, ...sesSet })
-        : verifyAndSendEmailSes({ mailObject, ...sesSet });
-    })
-    .then((res) => {
-      console.log("send mail complete");
-      return res;
-    })
-    .catch((err) => {
-      console.error(err);
-      const awsError = {
-        statusCode: 500,
-        body: {
-          errorMessage:
-            "メール検証済みチェックに問題がありました。Lambdaのログを確認してください。",
-        },
-      };
-      const resJson = JSON.stringify(awsError);
-      throw new Error(resJson);
-    });
-};
-
 export const handler = (event, context, callback) => {
   const eventClone = event;
   dotenv.config();
@@ -176,6 +61,14 @@ export const handler = (event, context, callback) => {
     mailAddress: emailAdmin,
   };
 
+  const verifyEmailIdentityCommand =
+    createVerifyEmailIdentityCommand(emailAdmin);
+  const sendEmailCommand = createSendEmailCommand(
+    emailAdmin,
+    emailAdmin,
+    eventClone
+  );
+
   return Promise.all([
     checkInvalidMailAddressFormat(emailAdmin),
     checkVoidString(eventClone.html),
@@ -187,6 +80,8 @@ export const handler = (event, context, callback) => {
       console.log(data);
       return confirmVerifiedAndSendEmailSes({
         mailObject: eventClone,
+        verifyEmailIdentityCommand,
+        sendEmailCommand,
         ...sesSet,
       });
     })
